requirements.txt
# Core Dependencies
apache-airflow==2.8.0
pandas>=2.0.0
numpy>=1.24.0

# Excel Support
openpyxl>=3.1.0

# Email Support (included in Python standard library)
# smtplib (built-in)
# email (built-in)

# Environment Variables
python-dotenv>=1.0.0

# Database Support
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# Data Validation
pydantic>=2.0.0

# YAML Support
PyYAML>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.0
pendulum>=2.1.2
